backend/requirements.txt
# Python dependencies
fastapi
uvicorn
pydantic
pydantic-ai
python-dotenv
